video/fbcon: Fix double buffering with non-fullscreen resolution
If an exact resolution match is not found, SDL will accept a higher
resolution and add black bars around it. For example, when 320x200 is
requested by only 320x240 is available, it will add 20 black lines on
the top and the bottom of the screen. The surface that SDL returns to
the application will look like it actually got 320x200.

When double buffering, the actual video mode's height should be used
instead of the height before adding the black borders. For the pixel
pointer this was done correctly, but for the vertical panning offset
the wrong height was used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>